move time event accumulator to from engine file to its own file
