+++ crates/backtest/src/lib.rs
@@ -41,6 +41,7 @@ pub mod exchange;
 pub mod execution_client;
 pub mod modules;
 pub mod runner;
+pub mod accumulator;
 
 #[cfg(feature = "ffi")]
 pub mod ffi;
